+++ src/ApisHandler.js
@@ -1,159 +1,84 @@
 "use strict";
 
-import { each, isArray } from "lodash";
+import { sources } from "@xbyorange/mercury";
 
-import { defaultConfig } from "./defaultConfig";
+export const TAG = "api";
+
+const SET_HEADERS_METHOD = "setHeaders";
+const ADD_HEADERS_METHOD = "addHeaders";
 
 export class ApisHandler {
-  constructor() {
-    this.reset();
-  }
-
   _ensureArray(tags) {
-    if (!tags) {
-      return [];
-    }
-    return isArray(tags) ? tags : [tags];
-  }
-
-  _setBaseConfig(config) {
-    this._baseConfig = { ...this._baseConfig, ...config };
-    each(this._all, api => {
-      api.config(this._baseConfig);
-    });
-  }
-
-  _setTagConfig(config, tag) {
-    this._configs[tag] = { ...this._configs[tag], ...config };
-    each(this._tagged[tag], api => {
-      api.config(this._configs[tag]);
-    });
-  }
-
-  _setBaseHeaders(headers) {
-    this._baseHeaders = headers;
-    each(this._all, api => {
-      api.setHeaders(this._baseHeaders);
-    });
-  }
-
-  _setTagHeaders(headers, tag) {
-    this._headers[tag] = headers;
-    each(this._tagged[tag], api => {
-      api.setHeaders(this._headers[tag]);
-    });
-  }
-
-  _addBaseHeaders(headers) {
-    this._baseAddedHeaders = { ...this._baseAddedHeaders, ...headers };
-    each(this._all, api => {
-      api.addHeaders(this._baseAddedHeaders);
-    });
-  }
-
-  _addTagHeaders(headers, tag) {
-    this._addedHeaders[tag] = { ...this._addedHeaders[tag], ...headers };
-    each(this._tagged[tag], api => {
-      api.addHeaders(this._addedHeaders[tag]);
-    });
-  }
-
-  _cleanAll() {
-    each(this._all, api => {
-      api.clean();
-    });
-  }
-
-  _clean(tag) {
-    each(this._tagged[tag], api => {
-      api.clean();
-    });
-  }
-
-  register(api, tags) {
-    this._all.push(api);
-    this._ensureArray(tags).forEach(tag => {
-      this._tagged[tag] = this._tagged[tag] || [];
-      this._tagged[tag].push(api);
-    });
-  }
-
-  getConfig(customConfig) {
-    const clonedCustomConfig = { ...customConfig };
-    delete clonedCustomConfig.tags;
-    let totalConfig = { ...defaultConfig, ...this._baseConfig, ...clonedCustomConfig };
-    this._ensureArray(customConfig.tags).forEach(tag => {
-      totalConfig = {
-        ...totalConfig,
-        ...this._configs[tag]
-      };
-    });
-    return totalConfig;
-  }
-
-  getHeaders(tags) {
-    let headers = { ...this._baseHeaders, ...this._baseAddedHeaders };
-    this._ensureArray(tags).forEach(tag => {
-      headers = {
-        ...headers,
-        ...this._headers[tag],
-        ...this._addedHeaders[tag]
-      };
-    });
-    return headers;
+    return Array.isArray(tags) ? tags : [tags];
   }
 
   config(configuration, tags) {
-    const config = { ...configuration };
+    console.warn(
+      'mercury-api apis.config method will be deprecated. Use mercury sources.getByTag("api").config instead'
+    );
     if (!tags) {
-      this._setBaseConfig(config);
+      sources.getByTag(TAG).config(configuration);
     } else {
       this._ensureArray(tags).forEach(tag => {
-        this._setTagConfig(config, tag);
+        sources.getByTag(tag).config(configuration);
       });
     }
   }
 
+  _getHeaders(tags) {
+    let headers = sources.getByTag(TAG)._apiHeaders;
+    tags.forEach(tag => {
+      const byTagSources = sources.getByTag(tag);
+      headers = { ...headers, ...byTagSources._apiHeaders };
+    });
+    return headers;
+  }
+
   setHeaders(headers, tags) {
-    const clonedHeaders = { ...headers };
     if (!tags) {
-      this._setBaseHeaders(clonedHeaders);
+      const allSources = sources.getByTag(TAG);
+      allSources._apiHeaders = headers;
+      allSources.call(SET_HEADERS_METHOD, headers);
     } else {
       this._ensureArray(tags).forEach(tag => {
-        this._setTagHeaders(clonedHeaders, tag);
+        const byTagSources = sources.getByTag(tag);
+        byTagSources._apiHeaders = headers;
+        sources.getByTag(tag).call(SET_HEADERS_METHOD, headers);
       });
     }
   }
 
   addHeaders(headers, tags) {
-    const clonedHeaders = { ...headers };
     if (!tags) {
-      this._addBaseHeaders(clonedHeaders);
+      const allSources = sources.getByTag(TAG);
+      allSources._apiHeaders = { ...allSources._apiHeaders, ...headers };
+      allSources.call(ADD_HEADERS_METHOD, headers);
     } else {
       this._ensureArray(tags).forEach(tag => {
-        this._addTagHeaders(clonedHeaders, tag);
+        const byTagSources = sources.getByTag(tag);
+        byTagSources._apiHeaders = { ...byTagSources._apiHeaders, ...headers };
+        sources.getByTag(tag).call(ADD_HEADERS_METHOD, headers);
       });
     }
   }
 
   clean(tags) {
+    console.warn(
+      `mercury-api apis.clean method will be deprecated. Use mercury sources.getByTag("api").clean instead`
+    );
     if (!tags) {
-      this._cleanAll();
+      sources.getByTag(TAG).clean();
     } else {
       this._ensureArray(tags).forEach(tag => {
-        this._clean(tag);
+        sources.getByTag(tag).clean();
       });
     }
   }
 
   reset() {
-    this._all = [];
-    this._tagged = {};
-    this._baseConfig = {};
-    this._configs = {};
-    this._baseHeaders = {};
-    this._headers = {};
-    this._baseAddedHeaders = {};
-    this._addedHeaders = {};
+    console.warn(
+      'mercury-api apis.reset method will be deprecated. Use mercury sources.getByTag("api").clean instead'
+    );
+    sources.getByTag(TAG).clear();
   }
 }